create script archive
